requirements.txt
yt_dlp==2025.5.22
aiogram==3.20.0.post0
python-dotenv==1.1.0